move consent back to xbr
