+++ cfxdb/xbrmm/schema.py
@@ -5,7 +5,7 @@
 #
 ##############################################################################
 
-from cfxdb.xbrmm.consent import Consents, IndexConsentByMemberAddress
+from cfxdb.xbr.consent import Consents, IndexConsentByMemberAddress
 
 from cfxdb.xbrmm.channel import PaymentChannels, IndexPaymentChannelByDelegate, \
     PaymentChannelBalances, PayingChannels, IndexPayingChannelByDelegate, \
